[UserCenter] about-us: move AboutUs to subdir
